+++ src/main/java/com/example/hcbc/domain/user/dto/ProfileResponseDto.java
@@ -1,5 +1,5 @@
 // src/main/java/com/example/hcbc/dto/ProfileResponse.java
-package com.example.hcbc.dto;
+package com.example.hcbc.domain.user.dto;
 
 import com.example.hcbc.global.common.enums.Category;
 import com.example.hcbc.global.common.enums.Gender;
