+++ src/main/java/com/example/hcbc/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.hcbc.repository;
+package com.example.hcbc.domain.user.repository;
 
-import com.example.hcbc.entity.User;
+import com.example.hcbc.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
